Fix #4503 (#4698)
* Fix #4503

- added check to make sure node exists before tying to get descendents and delete
- added test case that calls deleteNode on an undefined node

* formatting

* formatting

+++ packages/gatsby/src/redux/actions.js
@@ -412,13 +412,17 @@ actions.createLayout = (
  * deleteNode(node.id, node)
  */
 actions.deleteNode = (nodeId: string, node: any, plugin: Plugin) => {
-  // Also delete any nodes transformed from this one.
   let deleteDescendantsActions
-  const descendantNodes = findChildrenRecursively(node.children)
-  if (descendantNodes.length > 0) {
-    deleteDescendantsActions = descendantNodes.map(n =>
-      actions.deleteNode(n, getNode(n), plugin)
-    )
+  // It's possible the file node was never created as sometimes tools will
+  // write and then immediately delete temporary files to the file system.
+  if (node) {
+    // Also delete any nodes transformed from this one.
+    const descendantNodes = findChildrenRecursively(node.children)
+    if (descendantNodes.length > 0) {
+      deleteDescendantsActions = descendantNodes.map(n =>
+        actions.deleteNode(n, getNode(n), plugin)
+      )
+    }
   }
 
   const deleteAction = {
@@ -970,8 +974,8 @@ actions.createRedirect = ({
 }
 
 /**
-  * All defined actions.
-  */
+ * All defined actions.
+ */
 exports.actions = actions
 
 /**
